Add test for macros, fix resulting bug
In the bug, quasiquote would modify lists rather than producing new
ones. In this commit, we add a fix which creates a new list as we
quasiquote, rather than modifying the one passed in. This is more
expensive. Optimally, we'd only re-allocate nodes we have to change, but
this is the good enough solution.